+++ src/hash-table-helpers.js
@@ -1,4 +1,77 @@
+/* eslint-disable */
 // A special array class that can only store the number of items specified by the `limit` argument
+class LinkedList {
+  constructor() {
+    this.head = null;
+    this.tail = null;
+    // Do not modify anything inside of the constructor
+  }
+  // Wraps the given value in a node object and adds the node to the tail of the list
+  // If the list is empty, the new element is considered the tail as well as the head
+  // If there is one element in the list before the new element is added, the new element becomes the tail of the list
+  addToTail(values) {
+    const newNode = {
+      next: null,
+      value: values,
+    };
+    if (this.head === null) {
+      this.head = newNode;
+      this.tail = newNode;
+      return;
+    }
+
+    this.tail.next = newNode; // set the current tail node to point to the newNode
+    this.tail = newNode; // set the current tail to BE the newNode
+  }
+  // Removes the current head node from the list, replacing it with the next element in the list
+  // Returns the value of the removed node
+  removeHead() {
+    // need to check if there is a head node
+    if (this.head === null) return;
+    // check if head has a next
+    if (this.head.next === null) {
+      const head = this.head;
+      this.head = null;
+      this.tail = null;
+      return head.value;
+    }
+    const head = this.head;
+    this.head = this.head.next;
+    return head.value;
+  }
+  // Checks the linked list for the given value
+  // Returns true if the the value is found in the list, false otherwise
+  contains(value) {
+    // start at the head
+    let node = this.head;
+    while (node !== null) {
+      if (node.value[0] === value) {
+        return node.value[1];
+      }
+      node = node.next;
+    }
+    return false;
+  }
+  // replace the value
+  replace(value) {
+    // start at the head
+    let node = this.head;
+    while (node !== null) {
+      if (node.value[0] === value[0]) {
+        node.value[1] = value[1];
+        return;
+      }
+      node = node.next;
+    }
+    this.addToTail(value); // fix this
+    return;
+  }
+  // remove the value
+  removeSpot() {
+    
+  }
+}
+
 class LimitedArray {
   constructor(limit) {
     // You should not be directly accessing this array from your hash table methods
@@ -52,4 +125,5 @@ const getIndexBelowMax = (str, max) => {
 module.exports = {
   LimitedArray,
   getIndexBelowMax,
+  LinkedList,
 };
